requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
greenlet==3.1.1

# Redis
redis==5.2.1

# HTTP client
httpx==0.28.1

# Configuration
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Form data support
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]>=2.0
